Revert "x86, reboot: Use NMI instead of REBOOT_VECTOR to stop cpus"
upstream inclusion
from: mainline 3.5
target: kernel 3.4
category: bugfix

-----------------------------------------------------------

commit 5d2b86d90f7cc4a41316cef3d41560da6141f45c upstream.

This reverts commit 3603a2512f9e69dc87914ba922eb4a0812b21cd6.

Originally I wanted a better hammer to shutdown cpus during
panic. However, this really steps on the toes of various
spinlocks in the panic path.  Sometimes it is easier to wait for
the IRQ to become re-enabled to indictate the cpu left the
critical region and then shutdown the cpu.

The next patch moves the NMI addition after the IRQ part.  To
make it easier to see the logic of everything, revert this patch
and apply the next simpler patch.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/1336761675-24296-2-git-send-email-dzickus@redhat.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Integrated-by: Yijing Wang <wangyijing@huawei.com>