fc class: fix scanning when devs are offline
mainline inclusion
from mainline-3.5
upstream commit 8fb2ef89b696624ea9853868e370d2e132e76f54
category: bugfix

-----------------------------------------------------------

When a rport is added back or the role is changed the fc class
will queue a scan and then call scsi_target_unblock. The problem
with this is if the devices are in the SDEV_OFFLINE state and
the scan is run before the scsi_target_unblock, then the scan
will see LUN0 as offline and the scan will fail. This patch moves
the unblock call to before the scan, so we know the device state
will be set correctly when the scan is run.

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Integrated-by: Libo Chen <libo.chen@huawei.com>