block: restructure get_request()
mainline inclusion
from mainline-3.5
upstream commit b679281a6410676a41b175c5a185150a1ae42f9d
category: feature
feature: a mechanism to associate a bio to a task

-----------------------------

get_request() is structured a bit unusually in that failure path is
inlined in the usual flow with goto labels atop and inside it.
Relocate the error path to the end of the function.

This is to prepare for icq handling changes in get_request() and
doesn't introduce any behavior change.

Signed-off-by: Tejun Heo <tj@kernel.org>
Cc: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Integrated-by: Hans Zhang <zhanghonghui@huawei.com>