x86/mce: Avoid reading every machine check bank register twice.
mainline inclusion
from mainline-3.5
upstream commit 95022b8cf6ed7f3292b60c8e85fe59a12bfb1c9e
category: feature
feature: MCA recovery enhancement

suse inclusion information:
References: fate#313743, fate#314744
Target: SLE-11 SP3

This patchset enhances some MCA recovery features:
 - support recovery from instruction fetch error;
 - quirk recovery from instruction error on Sandy Bridge processors.

-----------------------------------------

Reading machine check bank registers is slow. There is a trend of
increasing the number of banks, and the number of cores. The main section
of do_machine_check() is a serialized section where each cpu in turn
checks every bank. Even on a little two socket SandyBridge-EP system
that multiplies out as:

	2 sockets * 8 cores * 2 hyperthreads * 20 banks = 640 MSRs

We already scan the banks in parallel in mce_no_way_out() to see if there
is a fatal error anywhere in the system. If we build a cache of VALID
bits during this scan, we can avoid uselessly re-reading banks that have
no data. Note that this cache is only a hint. If the valid bit is set in a
shared bank, all cpus that share that bank will see it during the parallel
scan, but the first to find it in the sequential scan will (usually) clear
the bank.

Acked-by: Borislav Petkov <borislav.petkov@amd.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>
Integrated-by: Xie XiuQi <xiexiuqi@huawei.com>