inotify: dont skip removal of watch descriptor if creation of ignored event failed
mainline inclusion
from mainline-3.8
upstream commit 8b99c3ccf735a2294c7842d236caa42e543e2c95
category: bugfix

-----------------------------------------------

In inotify_ignored_and_remove_idr() the removal of a watch descriptor is skipped
if the allocation of an ignored event failed and we are leaking memory (the
watch descriptor and the mark linked to it).
This patch ensures that the watch descriptor is removed regardless of whether
event creation failed or not.

Signed-off-by: Lino Sanfilippo <LinoSanfilippo@gmx.de>
Signed-off-by: Eric Paris <eparis@redhat.com>
Integrated-by: Rui Xiang <rui.xiang@huawei.com>