md/raid5: Allow reshape while a bitmap is present.
mainline inclusion
from mainline-3.5
upstream commit 30b67645faadcdc32072f999796b99bd0a20f2d5
category: bugfix

--------------------------------------------------------------------

We always should have allowed this.  A raid5 reshape doesn't change
the size of the bitmap, so not need to restrict it.

Also add a test to make sure we don't try to start a reshape on a
failed array.

Signed-off-by: NeilBrown <neilb@suse.de>
Integrated-by: Libo Chen <libo.chen@huawei.com>