DM RAID: Record and handle missing devices
mainline inclusion
from mainline-3.5
upstream commit 81f382f9e0b25ef56b1c0283c900b86b91a5e4c7
category: bugfix

--------------------------------------------------------------------

Missing dm-raid devices should be recorded in the superblock

When specifying the devices that compose a DM RAID array, it is possible to denote
failed or missing devices with '-'s.  When this occurs, we must record this in the
superblock.  We do this by checking if the array position's data device is missing
and then forcing MD to record the superblock by setting 'MD_CHANGE_DEVS' in
'raid_resume'.  If we do not cause the superblock to be rewritten by the resume
function, it is possible for a stale superblock to be written by an out-going
in-active table (during 'raid_dtr').

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Integrated-by: Libo Chen <libo.chen@huawei.com>