NFSv2/v3: Remove incorrect dprintks from the readdir reply code
mainline inclusion
from mainline-3.6
upstream commit db3a3bcf08d5c79290b369d70cc0b80c369d8ad9
category: bugfix

-------------------------------

The actual size of the directory is unknown to the client, so it is
always requesting the maximum number it can handle. If the server
is replying with fewer entries than was requested, then that will
usually reflect the fact that we've hit the end of the directory.
Flagging it as an error is therefore incorrect.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Integrated-by: Zhao Hongjiang <zhaohongjiang@huawei.com>