cls_cgroup: Allow classifier cgroups to have their classid reset to 0
mainline inclusion
from mainline-3.7
upstream commit 3afa6d00fb4f9712fbb44b63ba31f88b6f9239fe
category: bugfix

------------------------------------------------------------

The network classifier cgroup initalizes each cgroups instance classid value to
0.  However, the sock_update_classid function only updates classid's in sockets
if the tasks cgroup classid is not zero, and if it differs from the current
classid.  The later check is to prevent cache line dirtying, but the former is
detrimental, as it prevents resetting a classid for a cgroup to 0.  While this
is not a common action, it has administrative usefulness (if the admin wants to
disable classification of a certain group temporarily for instance).

Easy fix, just remove the zero check.  Tested successfully by myself

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
CC: "David S. Miller" <davem@davemloft.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
Integrated-by: Ding Tianhong <dingtianhong@huawei.com>