DM RAID: Fix for "sync" directive ineffectiveness
mainline inclusion
from mainline-3.7
upstream commit 761becff016b82a6a7a1b2ef224248da5f46bae9
category: bugfix

--------------------------------------------------------------------

There are two table arguments that can be given to a DM RAID target
that control whether the array is forced to (re)synchronize or skip
initialization: "sync" and "nosync".  When "sync" is given, we set
mddev->recovery_cp to 0 in order to cause the device to resynchronize.
This is insufficient if there is a bitmap in use, because the array
will simply look at the bitmap and see that there is no recovery
necessary.

The fix is to skip over the loading of the superblocks when "sync" is
given, causing new superblocks to be written that will force the array
to go through initialization (i.e. synchronization).

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Integrated-by: Libo Chen <libo.chen@huawei.com>