md: make sure manual changes to recovery checkpoint are saved.
mainline inclusion
from mainline-3.7
upstream commit db07d85ef6395d4768c6cdceeb2ea4878bd2a3f4
category: bugfix

--------------------------------------------------------------------

If you make an array bigger but suppress resync of the new region with
  mdadm --grow /dev/mdX --size=max --assume-clean

then stop the array before anything is written to it, the effect of
the "--assume-clean" is lost and the array will resync the new space
when restarted.
So ensure that we update the metadata in the case.

Reported-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Integrated-by: Libo Chen <libo.chen@huawei.com>