Doc: add Euler-Kernel-Submitting-Patches.txt
euler inclusion
category: doc

we need to document our rules of submitting patches to euler kernel,
it contains the basic step of submitting patches, and unified patch
format.

This guideline is so important for our euler kernel daily works,
so I prefer to place it into euler kernel root directory, not word
doc in SVN server, then it would be easy to access and reference it.

Signed-off-by: zhangwei(Jovi) <jovi.zhangwei@huawei.com>