netfilter: xt_TCPMSS: fix handling of malformed TCP header and options
mainline inclusion
from: v3.11
target: kernel 3.4
category: CVE-2017-18017
补丁描述: CVE-2017-18017的依赖补丁

----------------------

commit 71ffe9c77dd7a2b62207953091efa8dafec958dd upstream.

Make sure the packet has enough room for the TCP header and
that it is not malformed.

While at it, store tcph->doff*4 in a variable, as it is used
several times.

This patch also fixes a possible off by one in case of malformed
TCP options.

Reported-by: Julian Anastasov <ja@ssi.bg>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
Signed-off-by: Jason Yan <yanaijie@huawei.com>